+++ UIKit-Randomium.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "self:/Users/ivan-kalabishka/Documents/Development/SwiftBook/Module 2/Lesson 10/&#x41f;&#x43e;&#x432;&#x442;&#x43e;&#x440; &#x443;&#x440;&#x43e;&#x43a;&#x430;/&#x414;&#x417;/UIKit-Randomium/UIKit-Randomium.xcodeproj">
+   </FileRef>
+</Workspace>
