+++ UIKit-Randomium/Models/MusicAlbum.swift
@@ -1,0 +1,8 @@
+//
+//  MusicAlbum.swift
+//  UIKit-RandomCinema
+//
+//  Created by Kalabishka Ivan on 13.10.2021.
+//
+
+import Foundation
