+++ UIKit-Randomium/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
