+++ UIKit-Randomium/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
