+++ UIKit-Randomium/Resources/Assets.xcassets/Contents.json
