+++ UIKit-Randomium/ViewControllers/MainViewController.swift
@@ -1,5 +1,5 @@
 //
-//  FirstViewController.swift
+//  MainViewController.swift
 //  UIKit-RandomCinema
 //
 //  Created by Kalabishka Ivan on 06.10.2021.
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class FirstViewController: UIViewController {
+class MainViewController: UIViewController {
     
 }
